+++ src/main/java/com/clearbases/codehouse/dao/ProductDAO.java
@@ -26,7 +26,7 @@ public class ProductDAO {
         return entityManager.createQuery("SELECT p FROM Product p", Product.class).getResultList();
     }
 
-    public Product find(Long id) {
+    public Product find(Integer id) {
         return entityManager.createQuery("SELECT p FROM Product p join fetch p.prices WHERE p.id = :id", Product.class)
                 .setParameter("id", id)
                 .getSingleResult();
